% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabTNSData.R
\docType{data}
\name{Basal_Codes}
\alias{Basal_Codes}
\title{Basal_Codes}
\format{
A data frame.
}
\source{
Internal data.
}
\usage{
Basal_Codes
}
\description{
Basal codes used for frailty/comorbidity mapping.
}
\keyword{datasets}
