% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabTNSData.R
\docType{data}
\name{Code_Categories}
\alias{Code_Categories}
\title{Code_Categories}
\format{
A data frame.
}
\source{
Internal data.
}
\usage{
Code_Categories
}
\description{
Category definitions used in the comorbidity/frailty algorithms.
}
\keyword{datasets}
