% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabTNSData.R
\docType{data}
\name{Frailty_Comorbidity_Mapping}
\alias{Frailty_Comorbidity_Mapping}
\title{Frailty_Comorbidity_Mapping}
\format{
A data frame.
}
\source{
Internal data.
}
\usage{
Frailty_Comorbidity_Mapping
}
\description{
Mapping table between frailty items and comorbidity categories.
}
\keyword{datasets}
