% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabTNSData.R
\docType{data}
\name{Frailty_ICD10CM}
\alias{Frailty_ICD10CM}
\title{Frailty_ICD10CM}
\format{
A data frame.
}
\source{
Internal data.
}
\usage{
Frailty_ICD10CM
}
\description{
ICD-10-CM code lists used to compute frailty scores.
}
\keyword{datasets}
