% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabTNSData.R
\docType{data}
\name{Frailty_ICD11}
\alias{Frailty_ICD11}
\title{Frailty_ICD11}
\format{
A data frame.
}
\source{
Internal data.
}
\usage{
Frailty_ICD11
}
\description{
ICD-11 code lists used to compute frailty scores.
}
\keyword{datasets}
