% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabTNSData.R
\docType{data}
\name{Frailty_mapping}
\alias{Frailty_mapping}
\title{Frailty_mapping}
\format{
A data frame.
}
\source{
Internal data.
}
\usage{
Frailty_mapping
}
\description{
Mapping table used to compute frailty scores.
}
\keyword{datasets}
