% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coding_system.R
\name{get_coding_system}
\alias{get_coding_system}
\title{Choose a coding system}
\usage{
get_coding_system(
  coding_system = NULL,
  valid_systems = c("ICD-10-CA", "ICD-10-CM", "ICD-11"),
  verbose = interactive()
)
}
\arguments{
\item{coding_system}{Character string. If NULL and interactive(), the user is prompted.}

\item{valid_systems}{Character vector of allowed coding systems.}

\item{verbose}{Logical; if TRUE, prints guidance using message().}
}
\value{
A single character string indicating the selected coding system.
}
\description{
Choose a coding system
}
