% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabTNSData.R
\docType{data}
\name{ICD10CM_Basal_Codes}
\alias{ICD10CM_Basal_Codes}
\title{ICD10CM_Basal_Codes}
\format{
A data frame.
}
\source{
Internal data.
}
\usage{
ICD10CM_Basal_Codes
}
\description{
Basal codes for ICD-10-CM used by the package.
}
\keyword{datasets}
