% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabTNSData.R
\docType{data}
\name{ICD10CM_categorisation}
\alias{ICD10CM_categorisation}
\title{ICD10CM_categorisation}
\format{
A data frame.
}
\source{
Internal data.
}
\usage{
ICD10CM_categorisation
}
\description{
ICD-10-CM categorisation table used by the package.
}
\keyword{datasets}
