% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabTNSData.R
\docType{data}
\name{ICD11_Basal_Codes}
\alias{ICD11_Basal_Codes}
\title{ICD11_Basal_Codes}
\format{
A data frame.
}
\source{
Internal data.
}
\usage{
ICD11_Basal_Codes
}
\description{
Basal codes for ICD-11 used by the package.
}
\keyword{datasets}
