% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabTNSData.R
\docType{data}
\name{ICD11_categorisation}
\alias{ICD11_categorisation}
\title{ICD11_categorisation}
\format{
A data frame.
}
\source{
Internal data.
}
\usage{
ICD11_categorisation
}
\description{
ICD-11 categorisation table used by the package.
}
\keyword{datasets}
