% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity-package.R
\docType{package}
\name{LABTNSCPSS-package}
\alias{LABTNSCPSS-package}
\title{LABTNSCPSS: Calculation of Comorbidity and Frailty Scores}
\description{
Computes comorbidity indices and combined frailty scores for multiple ICD coding systems, including ICD-10-CA, ICD-10-CM, and ICD-11. The package provides tools to preprocess episode data, map diagnosis codes to chronic categories, propagate conditions across episodes, and generate comorbidity and frailty measures. The methods implemented are original to this package and were developed by the authors for research applications; a manuscript describing the methodology is currently in preparation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bayaniazadeh/LabTNSCPSSPackage}
  \item \doi{10.1016/j.ijmedinf.2026.106709}
  \item Report bugs at \url{https://github.com/bayaniazadeh/LabTNSCPSSPackage/issues}
}

}
\author{
\strong{Maintainer}: Azadeh Bayani \email{azadeh.bayani@umontreal.ca} (\href{https://orcid.org/0009-0002-7707-9602}{ORCID})

Other contributors:
\itemize{
  \item Jean Noel Nikiema \email{jean.nikiema@umontreal.ca} [contributor]
  \item Michèle Bally \email{michele.bally.chum@ssss.gouv.qc.ca} [contributor]
}

}
\keyword{internal}
