% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\title{Pipe operator}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

