% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabTNsCPSSPackage.R, R/package.R
\docType{package}
\name{LABTNSCPSS}
\alias{LABTNSCPSS}
\title{LABTNSCPSS: Tools for comorbidity and frailty scoring}
\description{
This package provides tools for preprocessing patient data, managing chronic pathologies,
calculating frailty scores, and computing comorbidity indices (Charlson and Elixhauser).

Tools for calculating comorbidity indices and frailty scores from ICD coding systems
(ICD-10-CA, ICD-10-CM, ICD-11).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bayaniazadeh/LabTNSCPSSPackage}
  \item \doi{10.1016/j.ijmedinf.2026.106709}
  \item Report bugs at \url{https://github.com/bayaniazadeh/LabTNSCPSSPackage/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/bayaniazadeh/LabTNSCPSSPackage}
  \item \doi{10.1016/j.ijmedinf.2026.106709}
  \item Report bugs at \url{https://github.com/bayaniazadeh/LabTNSCPSSPackage/issues}
}

}
\author{
\strong{Maintainer}: Azadeh Bayani \email{azadeh.bayani@umontreal.ca} (\href{https://orcid.org/0009-0002-7707-9602}{ORCID})

Other contributors:
\itemize{
  \item Jean Noel Nikiema \email{jean.nikiema@umontreal.ca} [contributor]
  \item Michèle Bally \email{michele.bally.chum@ssss.gouv.qc.ca} [contributor]
}

}
